split rho_mpi func from main charge.cpp
