+++ source/module_elecstate/CMakeLists.txt
@@ -17,6 +17,7 @@ list(APPEND objects
     potentials/write_pot.cpp
     module_charge/charge.cpp
     module_charge/charge_init.cpp
+    module_charge/charge_mpi.cpp
     module_charge/charge_broyden.cpp
     module_charge/charge_extra.cpp
     module_charge/charge_mixing.cpp
